% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_boot_RRi_fit.R
\name{plot.boot_RRi_fit}
\alias{plot.boot_RRi_fit}
\title{Plot method for boot_RRi_fit objects}
\usage{
\method{plot}{boot_RRi_fit}(x, ...)
}
\arguments{
\item{x}{An object of class "boot_RRi_fit".}

\item{...}{Additional arguments (unused).}
}
\value{
A ggplot object with panel of density plots to visualize the
bootstrap distributions of the RRi model parameters.
}
\description{
Generates a panel of density plots to visualize the bootstrap distributions
of the RRi model parameters. The method converts the bootstrap results to
long format and creates one density plot per parameter.
}
