% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_boot_RRi_fit.R
\name{print.summary.boot_RRi_fit}
\alias{print.summary.boot_RRi_fit}
\title{Print summary of boot_RRi_fit objects}
\usage{
\method{print}{summary.boot_RRi_fit}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.boot_RRi_fit".}

\item{...}{Additional arguments.}
}
\description{
Prints a human-readable summary of the bootstrap RRi model parameter estimates,
including the mean, standard deviation, and selected quantiles for each parameter.
}
