% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_boot_RRi_fit.R
\name{print.boot_RRi_fit}
\alias{print.boot_RRi_fit}
\title{Print method for boot_RRi_fit objects}
\usage{
\method{print}{boot_RRi_fit}(x, ...)
}
\arguments{
\item{x}{An object of class "boot_RRi_fit".}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
A message (of class \code{NULL}) with the number of bootstrap replicates and a preview
of  the first 6 bootstrap samples. Additionally, returns the input object
invisibly.
}
\description{
Displays a concise summary of the bootstrap RRi model parameter estimates.
The printed output shows the number of bootstrap replicates and a preview
of the parameter estimates from the first few replicates.
}
