% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_RRi_fit.R
\name{print.summary.RRi_fit}
\alias{print.summary.RRi_fit}
\title{Print summary of RRi_fit objects}
\usage{
\method{print}{summary.RRi_fit}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.RRi_fit".}

\item{...}{Additional arguments.}
}
\value{
A message (of class \code{NULL}) with optimization method used, estimated parameters,
objective value from the Hubber loss algorithm, performance statistics of
the RRi model and the convergence code of the optimization method.
}
\description{
Print summary of RRi_fit objects
}
